moving intersection calc to foldbuilder
